Checking if the file path has cache busting
Netlify doesn't allow for `?` in filenames. Because the plugin names files from the name in the HTML the cached string on the end is being added to the filename.

Here I'm splitting the filename into an array wherever there's a `?` and only taking the first part. I think this is the best place to do so?